+++ plugins/08_BattleForAzeroth/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v132
+
+* Reorganise BFA achievements and add achievement ids to groups for better tracking.
+
 # v130
 
 * Bump toc to 120000